New io package available. The package will house all the input output helper functions. The package currently contains three new functions:
save_hpgl( ): to save the HPGL commands returned by a Chiplotle-HPGL object in a text file.

export( ): to convert a Chiplotle-HPGL object (or list of objects) into an image file.

view( ): to preview a Chiplotle-HPGL object.

This is a working but partial commit. The functions currently have hard coded variables. These must be made user settable in the config file. This implementation will come in a future commit.