This commit introduces an new class: Group. It is intended to deprecate the old Container class. Like the Container class, it always has an 'xy' position attribute. Like the Container class, Groups can be nested. However, Group behaves a bit differently in that the elements it contains do not know about their container. i.e., here there is only a one way link, from Group to its contained shapes. This allows us to have the same *instance* of a shape in several groups. This also means that shapes contained in a group will no longer know their absolute position! It is the responsibility of the Group to move its sub-shapes.
An important improvement is that the Group can contain *any* Chiplotle HPGL class, both primitives and compound. This makes the distinction between the old `Container` and `HPGLContainer` unnecessary. Thus, the HPGLContainer will also deprecate. 

Note that this is the first commit in the process of transitioning from Containers to Groups. It is now half baked and there may be bugs. 

See the example `group_example.py` under examples.