New logging feature available! (The tools.logtools houses tools for logging purposes.) Any function can now write/display its own log message by simplyapplying to it the @applyLog decorator (see hpgltools.inflate_hpgl_string( ) for the first and only instance of this currently). Once the decorator is applied, use the construct 'name_of_function.logger.log_message_type( )' to log a message. For example, if we apply the decorator to `inflate_hpgl_string` like so:
@applyLog
def inflate_hpgl_string(string):
   ...


Then we can call the logger from within the function like so:

@applyLog
def inflate_hpgl_string(string):
   ...

   if (error):
      inflate_hpgl_string.logger.warning('This is a warning message')


That's it.

All log messages are written to the $HOME/.chiplotle/session.log file. The logs are currently per-session only. Each time Chiplotle is run the file is rewritten. 

The config.py file now has a 'verbose' variable that can be set to True or False. 
When set to True, all log messages are displayed in the console in addition to being written to the log file.

* * *

This commit also remove spurious imports in VirtualSerialPort.