This commit marks the beginning of an important structural change. - _CompoundHPGL was replaced with three abstract classes:  _HPGLCompound, _HPGLCompoundShape, and _HPGLCompoundDecorator. Both _HPGLCompoundShape and _HPGLCompoundDecorator inherit from _HPGLCompound. The old _CompoundHPGL shapes (found directly under the hpgl/compound directory) are now _HPGLCompoundShape types. They are essentially the same as before, with the only difference that they no longer support a `pen` attribute! The pen as a fixed attribute of shapes has been deprecated in favor of the more general "decorator" pattern.
Because the number and type of "embellishing" shape attributes (such as pen) can be potentially infinite and diverse, the new _HPGLCompoundDecorators are now here to support the addition of potentially any type of aggregate functionality (decoration) added to any shape. 

A decorator is applied to a shape at instantiation time. e.g., to apply a pen decorator DPen to a given `shape` (so that the shape is painted with a specific pen) we would simply do the following:
>>> shape = Circle((0, 0), 1000)
>>> DPen(shape, 1)

The parameters to the DPen decorator are the _HPGLCompound object to be decorated and whatever parameters the decorator requires, in this case pen number.
Note that any number of shape decorators can be applied to a shape (or other decorators) recursively. e.g., suppose we had a DWiggle decorator that randomly wiggles a shape every time it is printed / formatted, we can then do this:

>>> Wiggle(DPen(shape, 1))

this would yield a "wiggly" `shape` drawn with pen 1.


The actual decorators are pending.
NOTE that things may have broken in this commit!