* New non-destructive 'offset' and 'rotate' transforms available.
Note that we now currently have *3* transformation models running in parallel!
1. The simple but limited 'offset', 'rotation' and 'pivot' attributes of _Shape, which allow us to apply these transformations non-destructively.
2. The destructive transformations under the 'geometry/destructive_transforms' folder. 
3. The non-destructive and more general transformations under the 'geometry/transforms' folder.

Which one will prevail? Should we have more than one? My current inclination is to keep 2 and 3. 
Maybe we can choose one, but each has it's pros and cons. The most important difference is the following:
Destructive transformations are applied immediately, to the Paths, and cannot be undone. The advantage here is that each transformation is applied once only. Formatting simply convert points to HPGL.
The non-destructive transformations are stored in the shapes and are applied at format time. This means that every time we call format, the transformations are computed. This means more time if format is called many times on shapes, but it also means one can remove transformations from shapes to which these where previously applied to. It also means that transformations can "float" on groups. 
Since destructive transformations must be applied to Paths (because Groups have no points) applying a destructive transformation on a Group effectively applies it to its Paths, which means one must be careful about sharing the same Path among different Groups, since transforming an instance on a Group A immediately transforms the instance under Group B.