* 'instantiate_virtual_plotter( )' is now imported automatically when running Chiplotle (or when calling `from chiplotle import *` from a script).
* Added aliases for Coordinate and CoordinateArray in main __init__.py:
Coord = Coordinate
CoordArray = CoordinateArray

For lazy typers, you can now just type Coord(1, 2) rather than Coordinate(1, 2) and CoordArray(1, 2) rather than CoordinateArray(1, 2). Note that this works only when running Chiplotle from the command line or when calling `from chiplotle imnport *` from within the Python interpreter. This is intended for quick live coding and NOT for scripted code... certainly not for production code as it can cause confusing and is less legible.