Changed layers interface and implementation in moving towards having a unique interface and a model with more conceptual integrity.
Layers were attributes of shapes: 
>>> c = circle(123)
>>> c.layer = 'layer_name'

Now, Layers are Group containers:
>>> c = circle(123)
>>> l = layer(c, 'layer_name')

This unifies the interface by always being of the form x(y), as it is for transforms, groups, locks, and formatting. It also makes the layering explicit in the structure hierarchy, and makes more sense in terms of the domain model... shapes don't have to know anything about layers.
The behavior is the same as before, in that layers that are closer to the leaves in the graph hierarchy take precedence and 'shadow' lower layers.